+++ Cargo.toml
@@ -15,10 +15,15 @@ edition = "2018"
 [badges]
 travis-ci = { repository = "pimeys/rust-web-push" }
 
+[features]
+default = ["isahc"]
+hyper-client = ["hyper", "hyper-tls"] #use features = ["hyper-client"], default-features = false for about 300kb size decrease.
+
 [dependencies]
 futures = "^0.3"
-hyper = {version = "^0.14", features = ["client", "http1"]}
-hyper-tls = "^0.5"
+hyper = { version = "^0.14", features = ["client", "http1"], optional = true }
+hyper-tls = { version = "^0.5", optional = true }
+isahc = { version = "^1.4.0", optional = true }
 http = "^0.2"
 serde = "^1.0"
 serde_json = "^1.0"
@@ -28,7 +33,7 @@ ece = "^2.1.0"
 native-tls = "^0.2"
 base64 = "^0.13"
 openssl = "^0.10"
-time = {version = "^0.2", features = ["std"]}
+time = { version = "^0.2", features = ["std"] }
 lazy_static = "^1.4"
 chrono = "^0.4"
 log = "^0.4"
@@ -36,4 +41,4 @@ log = "^0.4"
 [dev-dependencies]
 argparse = "^0.2"
 regex = "^1.5"
-tokio = { version = "^1.1", features = ["macros","rt-multi-thread"] }
+tokio = { version = "^1.1", features = ["macros", "rt-multi-thread"] }
