+++ src/clients/mod.rs
@@ -1,0 +1,7 @@
+pub mod request_builder;
+
+#[cfg(feature = "hyper-client")]
+pub mod hyper_client;
+
+#[cfg(not(feature = "hyper-client"))]
+pub mod isahc_client;
