+++ src/clients/request_builder.rs
@@ -1,8 +1,7 @@
 use crate::{error::WebPushError, message::WebPushMessage};
 use http::header::{CONTENT_ENCODING, CONTENT_LENGTH, CONTENT_TYPE};
-use hyper::{Body};
 use http::{Request, StatusCode};
-//TODO rename this mod or move it to client
+
 #[derive(Deserialize, Serialize, Debug, PartialEq)]
 struct ErrorInfo {
     code: u16,
@@ -11,7 +10,30 @@ struct ErrorInfo {
     message: String,
 }
 
-pub fn build_request(message: WebPushMessage) -> Request<Body> {
+/// Builds the request to send to the push service.
+///
+/// This function is generic over the request body, this means that you can swap out client implementations
+/// even if they use different body types.
+///
+/// #Example
+///
+/// ```no_run
+/// # use web_push::{SubscriptionInfo, WebPushMessageBuilder};
+/// let info = SubscriptionInfo::new(
+///  "http://google.com",
+///  "BLMbF9ffKBiWQLCKvTHb6LO8Nb6dcUh6TItC455vu2kElga6PQvUmaFyCdykxY2nOSSL3yKgfbmFLRTUaGv4yV8",
+///  "xS03Fi5ErfTNH_l9WHE9Ig",
+///  );
+///
+///  let mut builder = WebPushMessageBuilder::new(&info).unwrap();
+///
+///  //Build the request for isahc
+///  let request = build_request::<isahc::Body>(builder.build().unwrap());
+/// ```
+pub fn build_request<T>(message: WebPushMessage) -> Request<T>
+where
+    T: From<Vec<u8>> + From<&'static str>, //This bound can be reduced to a &[u8] instead of str if needed
+{
     let mut builder = Request::builder()
         .method("POST")
         .uri(message.endpoint)
@@ -34,6 +56,7 @@ pub fn build_request(message: WebPushMessage) -> Request<Body> {
     }
 }
 
+/// Parses the response from the push service, and will return `Err` if the request was bad.
 pub fn parse_response(response_status: StatusCode, body: Vec<u8>) -> Result<(), WebPushError> {
     match response_status {
         status if status.is_success() => Ok(()),
@@ -59,12 +82,11 @@ pub fn parse_response(response_status: StatusCode, body: Vec<u8>) -> Result<(), 
 
 #[cfg(test)]
 mod tests {
+    use crate::clients::request_builder::*;
     use crate::error::WebPushError;
     use crate::http_ece::ContentEncoding;
     use crate::message::{SubscriptionInfo, WebPushMessageBuilder};
-    use crate::services::autopush::*;
-    use hyper::StatusCode;
-    use hyper::Uri;
+    use http::Uri;
 
     #[test]
     fn builds_a_correct_request_with_empty_payload() {
@@ -78,7 +100,7 @@ mod tests {
 
         builder.set_ttl(420);
 
-        let request = build_request(builder.build().unwrap());
+        let request = build_request::<isahc::Body>(builder.build().unwrap());
         let ttl = request.headers().get("TTL").unwrap().to_str().unwrap();
         let expected_uri: Uri = "http://google.com".parse().unwrap();
 
@@ -98,7 +120,7 @@ mod tests {
 
         builder.set_payload(ContentEncoding::Aes128Gcm, "test".as_bytes());
 
-        let request = build_request(builder.build().unwrap());
+        let request = build_request::<isahc::Body>(builder.build().unwrap());
 
         let encoding = request.headers().get("Content-Encoding").unwrap().to_str().unwrap();
